+++ core/data/src/main/java/com/kolown/data/datasource/remote/AuthDataSource.kt
@@ -4,7 +4,7 @@ import androidx.credentials.CustomCredential
 import com.google.android.libraries.identity.googleid.GoogleIdTokenCredential
 import com.google.firebase.auth.FirebaseAuth
 import com.google.firebase.auth.GoogleAuthProvider
-import com.kolown.data.remote.UserDto
+import com.kolown.model.User
 import kotlinx.coroutines.tasks.await
 import javax.inject.Inject
 import javax.inject.Named
@@ -12,11 +12,11 @@ import javax.inject.Named
 interface AuthDataSource {
     suspend fun signInWithCredential(credential: CustomCredential): Result<Unit>
     fun getUserId(): String
-    fun getUserInfo(): UserDto
+    fun getUserInfo(): User
     fun checkUserLoggedIn(): Boolean
     fun logout(): Result<Unit>
-    suspend fun joinWithEmailAndPassword(email: String, password: String): Result<UserDto>
-    suspend fun signInWithEmailAndPassword(email: String, password: String): Result<UserDto>
+    suspend fun joinWithEmailAndPassword(email: String, password: String): Result<User>
+    suspend fun signInWithEmailAndPassword(email: String, password: String): Result<User>
 }
 
 @Named("google")
@@ -39,10 +39,10 @@ class AuthDataSourceImpl @Inject constructor(
         return "user-${currentUser.uid}"
     }
 
-    override fun getUserInfo(): UserDto {
+    override fun getUserInfo(): User {
         val currentUser = auth.currentUser ?: throw Exception("로그인 안된 유저")
 
-        return UserDto(userId = "user-${currentUser.uid}", email = currentUser.email.orEmpty())
+        return User(userId = "user-${currentUser.uid}", email = currentUser.email.orEmpty())
     }
 
     override fun checkUserLoggedIn(): Boolean {
@@ -58,10 +58,10 @@ class AuthDataSourceImpl @Inject constructor(
     override suspend fun signInWithEmailAndPassword(
         email: String,
         password: String,
-    ): Result<UserDto> {
+    ): Result<User> {
         return kotlin.runCatching {
             auth.signInWithEmailAndPassword(email, password).await().let { user ->
-                UserDto(
+                User(
                     userId = "user-${user.user?.uid}",
                     email = user.user?.email.orEmpty()
                 )
@@ -72,10 +72,10 @@ class AuthDataSourceImpl @Inject constructor(
     override suspend fun joinWithEmailAndPassword(
         email: String,
         password: String,
-    ): Result<UserDto> {
+    ): Result<User> {
         return kotlin.runCatching {
             auth.createUserWithEmailAndPassword(email, password).await().let {
-                UserDto(
+                User(
                     userId = "user-${it.user?.uid}",
                     email = it.user?.email.orEmpty()
                 )
